requirements.txt
Flask==3.0.0
openpyxl==3.1.2
gunicorn==21.2.0
python-dotenv==1.0.0
Werkzeug==3.1.3
Flask-WTF==1.2.2
